Work for Importing CSV into Stash
A bit of work that is used to load and manipulate a CSV file originally
from a side project of technologies that are worth investigating or
maybe will be.  CSV can be imported into memory (sotred in the resources
folder) parsed and used to create issues in a project I run.

Bit pointless for others to actually use my repo of course and I expect
most of this is useless for everyone else but there you go